move w sigil trailing comma test to warning_test.exs
